.claude/skills/social-meta/requirements.txt
# Facebook + Instagram Meta Skill Dependencies

requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
